Allow JSON requests to specify custom MIME types
Some JSON-based APIs have special behavior that can only be accessed by
specifying complex MIME types.  For example, if we want to ask
AllegroGraph for RDF context data in queries, we need to specify a MIME
type of 'application/x-quints+json'.

To support this, we allow the caller to override the default MIME values
used for a JSON request.